+++ lib/features/workout/screens/rest_screen.dart
@@ -10,9 +10,9 @@ import "package:pull_up_club/common/themes/app_typography.dart";
 import "package:pull_up_club/common/utils/utils.dart";
 import "package:pull_up_club/common/widgets/core/gradient_button.dart";
 import "package:pull_up_club/common/widgets/core/screen_scaffold.dart";
-import "package:pull_up_club/common/widgets/shared/set_cards.dart";
 import "package:pull_up_club/features/workout/providers/workout_provider.dart";
 import "package:pull_up_club/features/workout/widgets/destructive_workout_buttons.dart";
+import "package:pull_up_club/features/workout/widgets/previous_workout_set_cards.dart";
 
 class RestScreen extends StatefulWidget {
   const RestScreen({required this.numCompletedGroups, super.key});
@@ -94,9 +94,10 @@ class _RestScreenState extends State<RestScreen> {
               textColor: AppColors.skipRestText,
             ),
           ),
-          SetCards(
+          PreviousWorkoutSetCards(
             values: getSetCardValues(workout),
             numExpectedCards: workout.maxGroups,
+            workoutType: workout.workoutType,
             highlightedIndex: widget.numCompletedGroups,
           ),
           DestructiveWorkoutButtons(workout: workout, onUndoLastSet: _undoLastSet),